+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/config/PermissionFilter.java
@@ -4,10 +4,10 @@ import com.alibaba.fastjson.JSONObject;
 import lombok.extern.slf4j.Slf4j;
 import org.cloud.sonic.common.http.RespEnum;
 import org.cloud.sonic.common.http.RespModel;
-import org.cloud.sonic.common.models.domain.Resources;
-import org.cloud.sonic.common.services.ResourcesService;
-import org.cloud.sonic.common.services.RolesServices;
 import org.cloud.sonic.common.tools.JWTTokenTool;
+import org.cloud.sonic.controller.models.domain.Resources;
+import org.cloud.sonic.controller.services.ResourcesService;
+import org.cloud.sonic.controller.services.RolesServices;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
