+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/mapper/ResourcesMapper.java
@@ -2,8 +2,7 @@ package org.cloud.sonic.controller.mapper;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.apache.ibatis.annotations.Mapper;
-import org.cloud.sonic.common.models.domain.Agents;
-import org.cloud.sonic.common.models.domain.Resources;
+import org.cloud.sonic.controller.models.domain.Resources;
 
 /**
  *  Mapper 接口
