+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/mapper/RoleResourcesMapper.java
@@ -3,8 +3,7 @@ package org.cloud.sonic.controller.mapper;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
-import org.cloud.sonic.common.models.domain.Resources;
-import org.cloud.sonic.common.models.domain.RoleResources;
+import org.cloud.sonic.controller.models.domain.RoleResources;
 
 /**
  *  Mapper 接口
