+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/mapper/RolesMapper.java
@@ -2,8 +2,7 @@ package org.cloud.sonic.controller.mapper;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.apache.ibatis.annotations.Mapper;
-import org.cloud.sonic.common.models.domain.Resources;
-import org.cloud.sonic.common.models.domain.Roles;
+import org.cloud.sonic.controller.models.domain.Roles;
 
 /**
  *  Mapper 接口
