+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/models/domain/RoleResources.java
@@ -1,4 +1,5 @@
-package org.cloud.sonic.common.models.domain;
+package org.cloud.sonic.controller.models.domain;
+
 
 import com.baomidou.mybatisplus.annotation.*;
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -6,15 +7,12 @@ import com.gitee.sunchenbin.mybatis.actable.annotation.*;
 import com.gitee.sunchenbin.mybatis.actable.constants.MySqlCharsetConstant;
 import com.gitee.sunchenbin.mybatis.actable.constants.MySqlEngineConstant;
 import com.gitee.sunchenbin.mybatis.actable.constants.MySqlTypeConstant;
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
-import org.cloud.sonic.common.models.base.TypeConverter;
-import org.cloud.sonic.common.models.dto.ResourcesDTO;
 
 import java.io.Serializable;
 import java.util.Date;
@@ -24,41 +22,22 @@ import java.util.Date;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@TableName("resources")
-@TableComment("资源信息表")
+@TableName("resource_roles")
+@TableComment("角色资源表表")
 @TableCharset(MySqlCharsetConstant.DEFAULT)
 @TableEngine(MySqlEngineConstant.InnoDB)
-public class Resources implements Serializable, TypeConverter<Resources, ResourcesDTO> {
-
+public class RoleResources implements Serializable{
     @TableId(value = "id", type = IdType.AUTO)
     @IsAutoIncrement
     private Integer id;
 
-    @TableField("`desc`")
-    @Column(value = "desc", isNull = false, comment = "描述")
-    private String desc;
+    @TableField("role_id")
+    @Column(value = "role_id", isNull = false, comment = "描述")
+    private Integer roleId;
 
-    @TableField
-    @Column(value = "parent_id", isNull = false,defaultValue = "0", comment = "父级 id")
-    private Integer parentId;
-
-    @TableField
-    @Column(value = "method", isNull = false, comment = "请求方法")
-    @Index
-    private String method;
-
-    @TableField
-    @Column(value = "path", isNull = false, comment = "资源路径")
-    @Index
-    private String path;
-
-    @TableField
-    @Column(value = "white", isNull = false,defaultValue = "1", comment = "是否是白名单 url，0是 1 不是")
-    private Integer white;
-
-    @TableField
-    @Column(value = "need_auth", isNull = false,defaultValue = "1", comment = "是否需要鉴权，0 不需要 1 需要")
-    private Integer needAuth;
+    @TableField("res_id")
+    @Column(value = "res_id", isNull = false, comment = "描述")
+    private Integer resId;
 
     @ApiModelProperty(value = "创建时间", example = "2021-08-15 11:36:00")
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
@@ -71,6 +50,4 @@ public class Resources implements Serializable, TypeConverter<Resources, Resourc
     @TableField(fill = FieldFill.INSERT_UPDATE)
     @Column(value = "update_time", type = MySqlTypeConstant.DATETIME, comment = "更新时间")
     Date updateTime;
-
-
 }
