+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/models/dto/ResourcesDTO.java
@@ -1,4 +1,4 @@
-package org.cloud.sonic.common.models.dto;
+package org.cloud.sonic.controller.models.dto;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
@@ -7,11 +7,10 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
-import org.cloud.sonic.common.models.base.TypeConverter;
-import org.cloud.sonic.common.models.domain.Resources;
+import org.cloud.sonic.controller.models.base.TypeConverter;
+import org.cloud.sonic.controller.models.domain.Resources;
 
 import java.io.Serializable;
-import java.util.Date;
 import java.util.List;
 
 @ApiModel("请求资源DTO 模型")
