+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/models/dto/RolesDTO.java
@@ -1,4 +1,4 @@
-package org.cloud.sonic.common.models.dto;
+package org.cloud.sonic.controller.models.dto;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
@@ -7,8 +7,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
-import org.cloud.sonic.common.models.base.TypeConverter;
-import org.cloud.sonic.common.models.domain.Roles;
+import org.cloud.sonic.controller.models.base.TypeConverter;
+import org.cloud.sonic.controller.models.domain.Roles;
 
 import java.io.Serializable;
