+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/models/params/ResourceParam.java
@@ -1,4 +1,4 @@
-package org.cloud.sonic.common.models.params;
+package org.cloud.sonic.controller.models.params;
 
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
