+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/services/ResourcesService.java
@@ -1,10 +1,10 @@
-package org.cloud.sonic.common.services;
+package org.cloud.sonic.controller.services;
 
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.IService;
-import org.cloud.sonic.common.models.base.CommentPage;
-import org.cloud.sonic.common.models.domain.Resources;
-import org.cloud.sonic.common.models.dto.ResourcesDTO;
+import org.cloud.sonic.controller.models.base.CommentPage;
+import org.cloud.sonic.controller.models.domain.Resources;
+import org.cloud.sonic.controller.models.dto.ResourcesDTO;
 
 import java.util.List;
