+++ sonic-server-controller/src/main/java/org/cloud/sonic/controller/services/impl/ResourcesServiceImpl.java
@@ -6,17 +6,17 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
 import org.cloud.sonic.common.config.WhiteUrl;
-import org.cloud.sonic.common.models.base.CommentPage;
-import org.cloud.sonic.common.models.base.TypeConverter;
-import org.cloud.sonic.common.models.domain.Resources;
-import org.cloud.sonic.common.models.domain.RoleResources;
-import org.cloud.sonic.common.models.dto.ResourcesDTO;
-import org.cloud.sonic.common.models.interfaces.UrlType;
-import org.cloud.sonic.common.services.ResourcesService;
-import org.cloud.sonic.common.tools.SpringTool;
 import org.cloud.sonic.controller.mapper.ResourcesMapper;
 import org.cloud.sonic.controller.mapper.RoleResourcesMapper;
+import org.cloud.sonic.controller.models.base.CommentPage;
+import org.cloud.sonic.controller.models.base.TypeConverter;
+import org.cloud.sonic.controller.models.domain.Resources;
+import org.cloud.sonic.controller.models.domain.RoleResources;
+import org.cloud.sonic.controller.models.dto.ResourcesDTO;
+import org.cloud.sonic.controller.models.interfaces.UrlType;
+import org.cloud.sonic.controller.services.ResourcesService;
 import org.cloud.sonic.controller.services.impl.base.SonicServiceImpl;
+import org.cloud.sonic.controller.tools.SpringTool;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
